+++ modules/aws-iam-policies/main.tf
@@ -7,6 +7,8 @@ resource "aws_iam_policy" "emr_creator_minimal_policy" {
   policy = data.aws_iam_policy_document.emr_creator_policy.json
 }
 
+data "aws_caller_identity" "current" {}
+
 data "aws_iam_policy_document" "emr_creator_policy" {
   version = "2012-10-17"
   statement {
@@ -22,8 +24,8 @@ data "aws_iam_policy_document" "emr_creator_policy" {
       "iam:PassRole"
     ]
     resources = [
-      "arn:aws:iam::${var.aws_account_id}:role/*",
-      "arn:aws:elasticmapreduce:*:${var.aws_account_id}:cluster/*"
+      "arn:aws:iam::${data.aws_caller_identity.current.account_id}:role/*",
+      "arn:aws:elasticmapreduce:*:${data.aws_caller_identity.current.account_id}:cluster/*"
     ]
   }
   statement {
@@ -47,37 +49,12 @@ data "aws_iam_policy_document" "emr_creator_policy" {
 
 //Attach the above policy to an existing user
 resource "aws_iam_role_policy_attachment" "emr_creator_policy_attachment" {
-  role       = "${var.aws_role_name}"
-  policy_arn = "${aws_iam_policy.emr_creator_minimal_policy.arn}"
-}
-
-// reduced permissions policy for Tamr user to be able to access EMRFS
-resource "aws_iam_policy" "emrfs_user_minimal_policy" {
-  name   = var.aws_emrfs_user_policy_name
-  policy = data.aws_iam_policy_document.emrfs_user_policy.json
+  role       = var.aws_role_name
+  policy_arn = aws_iam_policy.emr_creator_minimal_policy.arn
 }
 
 data "aws_iam_policy_document" "emrfs_user_policy" {
   version = "2012-10-17"
-  statement {
-    effect = "Allow"
-    actions = [
-      "s3:PutObject",
-      "s3:GetObject",
-      "s3:DeleteObject",
-      "s3:AbortMultipartUpload",
-      "s3:ListBucket",
-      "s3:ListObjects"
-    ]
-    resources = [
-      "arn:aws:s3:::${var.aws_emrfs_hbase_bucket_name}/*",
-      "arn:aws:s3:::${var.aws_emrfs_hbase_bucket_name}",
-      "arn:aws:s3:::${var.aws_emrfs_hbase_logs_bucket_name}",
-      "arn:aws:s3:::${var.aws_emrfs_hbase_logs_bucket_name}/*",
-      "arn:aws:s3:::${var.aws_emrfs_spark_logs_bucket_name}",
-      "arn:aws:s3:::${var.aws_emrfs_spark_logs_bucket_name}/*"
-    ]
-  }
   statement {
     effect = "Allow"
     actions = [
@@ -88,8 +65,21 @@ data "aws_iam_policy_document" "emrfs_user_policy" {
   }
 }
 
+// reduced permissions policy for Tamr user to be able to access EMRFS
+resource "aws_iam_policy" "emrfs_user_minimal_policy" {
+  name   = var.aws_emrfs_user_policy_name
+  policy = data.aws_iam_policy_document.emrfs_user_policy.json
+}
+
+// IAM role policy attachment(s) that attach s3 policy ARNs to Tamr user IAM role
+resource "aws_iam_role_policy_attachment" "emrfs_user_s3_policy" {
+  count      = length(var.s3_policy_arns)
+  role       = var.aws_role_name
+  policy_arn = element(var.s3_policy_arns, count.index)
+}
+
 //Attach the above policy to an existing user
 resource "aws_iam_role_policy_attachment" "emrfs_user_policy_attachment" {
-  role       = "${var.aws_role_name}"
-  policy_arn = "${aws_iam_policy.emrfs_user_minimal_policy.arn}"
+  role       = var.aws_role_name
+  policy_arn = aws_iam_policy.emrfs_user_minimal_policy.arn
 }